+++ src/Type/HtmlType.php
@@ -10,10 +10,8 @@ declare(strict_types=1);
 namespace Ixocreate\Schema\Type;
 
 use Doctrine\DBAL\Types\JsonType;
-use nadar\quill\BlockListener;
-use nadar\quill\InlineListener;
-use nadar\quill\Lexer;
-use nadar\quill\Line;
+use Ixocreate\QuillRenderer\Renderer;
+use Ixocreate\Schema\Type\Html\IxoLink;
 
 final class HtmlType extends AbstractType implements DatabaseTypeInterface
 {
@@ -56,61 +54,11 @@ final class HtmlType extends AbstractType implements DatabaseTypeInterface
             return $this->value()['html'];
         }
 
-        $lexer = new Lexer($this->value()['quill']);
-        $lexer->escapeInput = true;
-        $lexer->registerListener(new class() extends InlineListener {
-            public function process(Line $line)
-            {
-                try {
-                    $link = $line->getAttribute('ixolink');
-                    if ($link) {
-                        /** @var LinkType $link */
-                        $link = Type::create($link, LinkType::serviceName());
-                        $this->updateInput(
-                            $line,
-                            '<a href="' . (string)$link . '" target="' . $link->getTarget() . '">' . $line->input . '</a>'
-                        );
-                    }
-                } catch (\Exception $exception) {
-                }
-            }
-        });
+        $renderer = new Renderer();
+        $renderer->enableDefaults();
+        $renderer->addInsert(new IxoLink());
 
-        $lexer->registerListener(new class() extends BlockListener {
-            /**
-             * @param Line $line
-             * @return void
-             */
-            public function process(Line $line)
-            {
-                $linebreak = $line->getAttribute('linebreak');
-                if ($linebreak === 'true') {
-                    $this->pick($line);
-                    $line->setDone();
-                }
-            }
-
-            public function render(Lexer $lexer)
-            {
-                foreach ($this->picks() as $pick) {
-                    // get all
-                    $prev = $pick->line->previous(function (Line $line) {
-                        if (!$line->isInline()) {
-                            return true;
-                        }
-                    });
-
-                    // if there is no previous element, we take the same line element.
-                    if (!$prev) {
-                        $prev = $pick->line;
-                    }
-
-                    $pick->line->output = $prev->input . $pick->line->renderPrepend() . '<br>';
-                    $prev->setDone();
-                }
-            }
-        });
-        return $lexer->render();
+        return $renderer->render($this->value()['quill']);
     }
 
     /**